Rename gazetteer build pipeline stages and reorganize related files
